relay stats test: put fixture into test file
